Packaged JSON schema validation methods as a new validation rule in the "all" category
